# src/aspnetcore/src/Org.OpenAPITools/Controllers/PluginsApi.cs
/*
 * SalesTim API
 *
 * The Microsoft Teams Governance API 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@salestim.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using Org.OpenAPITools.Attributes;
using Microsoft.AspNetCore.Authorization;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class PluginsApiController : ControllerBase
    { 
        /// <summary>
        /// Create a new plugin
        /// </summary>
        /// <remarks>Create a new plugin and get its generated ID and secret.</remarks>
        /// <param name="inlineObject1"></param>
        /// <response code="200">Operation executed successfully.</response>
        /// <response code="404">The ressource cannot be found.</response>
        /// <response code="429">Too many requests. Your request has been throttled.</response>
        /// <response code="401">API key or app id are missing or invalid.</response>
        /// <response code="500">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="0">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        [HttpPost]
        [Route("/v1.0/plugins")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("CreatePlugin")]
        [SwaggerResponse(statusCode: 200, type: typeof(Plugin), description: "Operation executed successfully.")]
        public virtual IActionResult CreatePlugin([FromBody]InlineObject1 inlineObject1)
        { 

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Plugin));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            //TODO: Uncomment the next line to return response 429 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(429);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "{\n  \"templateConfiguration\" : {\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"createdBy\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"name\" : \"name\",\n    \"description\" : \"description\",\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"secret\" : \"secret\",\n    \"enabled\" : true\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Plugin>(exampleJson)
            : default(Plugin);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete a plugin
        /// </summary>
        /// <remarks>Delete a plugin</remarks>
        /// <param name="id">ID of the plugin to delete</param>
        /// <response code="200">Operation executed successfully.</response>
        /// <response code="404">The ressource cannot be found.</response>
        /// <response code="429">Too many requests. Your request has been throttled.</response>
        /// <response code="401">API key or app id are missing or invalid.</response>
        /// <response code="500">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="0">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        [HttpDelete]
        [Route("/v1.0/plugins/{id}")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("DeletePlugin")]
        public virtual IActionResult DeletePlugin([FromRoute][Required]string id)
        { 

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            //TODO: Uncomment the next line to return response 429 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(429);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get all plugins in a tenant
        /// </summary>
        /// <remarks>Get all plugins in a tenant</remarks>
        /// <response code="200">Operation executed successfully.</response>
        /// <response code="404">The ressource cannot be found.</response>
        /// <response code="429">Too many requests. Your request has been throttled.</response>
        /// <response code="401">API key or app id are missing or invalid.</response>
        /// <response code="500">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="0">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        [HttpGet]
        [Route("/v1.0/plugins")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("GetPlugIns")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Plugin>), description: "Operation executed successfully.")]
        public virtual IActionResult GetPlugIns()
        { 

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Plugin>));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            //TODO: Uncomment the next line to return response 429 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(429);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "{\n  \"templateConfiguration\" : {\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"createdBy\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"name\" : \"name\",\n    \"description\" : \"description\",\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"secret\" : \"secret\",\n    \"enabled\" : true\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<Plugin>>(exampleJson)
            : default(List<Plugin>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Regenerate a plugin secret
        /// </summary>
        /// <remarks>Regenerate a plugin secret</remarks>
        /// <param name="id">The plugin ID</param>
        /// <response code="200">Operation executed successfully.</response>
        /// <response code="404">The ressource cannot be found.</response>
        /// <response code="429">Too many requests. Your request has been throttled.</response>
        /// <response code="401">API key or app id are missing or invalid.</response>
        /// <response code="500">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="0">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        [HttpPost]
        [Route("/v1.0/plugins/{id}/regenerateSecret")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("RegeneratePluginSecret")]
        [SwaggerResponse(statusCode: 200, type: typeof(Plugin), description: "Operation executed successfully.")]
        public virtual IActionResult RegeneratePluginSecret([FromRoute][Required]string id)
        { 

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Plugin));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            //TODO: Uncomment the next line to return response 429 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(429);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "{\n  \"templateConfiguration\" : {\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"createdBy\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"name\" : \"name\",\n    \"description\" : \"description\",\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"secret\" : \"secret\",\n    \"enabled\" : true\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Plugin>(exampleJson)
            : default(Plugin);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Update a plugin
        /// </summary>
        /// <remarks>Update a plugin</remarks>
        /// <param name="id">ID of the plugin to update</param>
        /// <param name="inlineObject2"></param>
        /// <response code="200">Operation executed successfully.</response>
        /// <response code="404">The ressource cannot be found.</response>
        /// <response code="429">Too many requests. Your request has been throttled.</response>
        /// <response code="401">API key or app id are missing or invalid.</response>
        /// <response code="500">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="0">The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        [HttpPut]
        [Route("/v1.0/plugins/{id}")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("UpdatePlugin")]
        [SwaggerResponse(statusCode: 200, type: typeof(Plugin), description: "Operation executed successfully.")]
        public virtual IActionResult UpdatePlugin([FromRoute][Required]string id, [FromBody]InlineObject2 inlineObject2)
        { 

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Plugin));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            //TODO: Uncomment the next line to return response 429 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(429);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "{\n  \"templateConfiguration\" : {\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"createdBy\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"name\" : \"name\",\n    \"description\" : \"description\",\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\n    \"secret\" : \"secret\",\n    \"enabled\" : true\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Plugin>(exampleJson)
            : default(Plugin);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}