# src/aspnetcore/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "contact" : {
      "email" : "support@salestim.com",
      "name" : "SalesTim Support",
      "url" : "https://developers.salestim.com/api"
    },
    "description" : "The Microsoft Teams Governance API\n",
    "license" : {
      "name" : "(c) Copyright 2019-2020 SalesTim - all rights reserved.",
      "url" : "https://www.salestim.com/legal/eula"
    },
    "termsOfService" : "https://www.salestim.com/legal/tos",
    "title" : "SalesTim API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "SalesTim API v1.0 (Production environment, uses live data)",
    "url" : "https://api.salestim.io/v1.0"
  } ],
...
  "paths" : {
...
    "/plugins" : {
      "get" : {
        "description" : "Get all plugins in a tenant",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins/#getPlugIns"
        },
        "operationId" : "GetPlugIns",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfPlugins"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Get all plugins in a tenant",
        "tags" : [ "Plugins" ]
      },
      "post" : {
        "description" : "Create a new plugin and get its generated ID and secret.",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins/#createPlugin"
        },
        "operationId" : "CreatePlugin",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/inline_object_1",
          "content" : {
            "application/json" : {
              "schema" : {
                "description" : "Object defining the properties to update and their new values",
                "properties" : {
                  "id" : {
                    "description" : "The template ID to be installed",
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "description" : "List of properties to be updated.",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Plugin"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Create a new plugin",
        "tags" : [ "Plugins" ]
      }
    },
    "/plugins/{id}/regenerateSecret" : {
      "post" : {
        "description" : "Regenerate a plugin secret",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins/#regeneratePluginSecret"
        },
        "operationId" : "RegeneratePluginSecret",
        "parameters" : [ {
          "description" : "The plugin ID",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Plugin"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Regenerate a plugin secret",
        "tags" : [ "Plugins" ]
      }
    },
    "/plugins/{id}" : {
      "delete" : {
        "description" : "Delete a plugin",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins/#deletePlugin"
        },
        "operationId" : "DeletePlugin",
        "parameters" : [ {
          "description" : "ID of the plugin to delete",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Delete a plugin",
        "tags" : [ "Plugins" ]
      },
      "put" : {
        "description" : "Update a plugin",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins/#updatePlugin"
        },
        "operationId" : "UpdatePlugin",
        "parameters" : [ {
          "description" : "ID of the plugin to update",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/inline_object_2",
          "content" : {
            "application/json" : {
              "schema" : {
                "description" : "Object defining the properties to update and their new values",
                "properties" : {
                  "name" : {
                    "description" : "The plugin name",
                    "type" : "string"
                  },
                  "description" : {
                    "description" : "The plugin description",
                    "type" : "string"
                  },
                  "enabled" : {
                    "description" : "The plugin status, enabled or disabled",
                    "type" : "boolean"
                  }
                },
                "type" : "object"
              }
            }
          },
          "description" : "List of properties to be updated.",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Plugin"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Update a plugin",
        "tags" : [ "Plugins" ]
      }
    },
...
  },
  "components" : {
    "requestBodies" : {
      "inline_object_1" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/inline_object_1"
            }
          }
        },
        "required" : true
      },
...
      "inline_object_2" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/inline_object_2"
            }
          }
        },
        "required" : true
      }
    },
...
    "schemas" : {
...
      "ArrayOfPlugins" : {
        "description" : "A list of platform plugins (aka \"Virtual Apps\")",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins"
        },
        "items" : {
          "$ref" : "#/components/schemas/Plugin"
        },
        "type" : "array"
      },
      "Plugin" : {
        "description" : "A platform plugin (aka \"Virtual Apps\")",
        "example" : {
          "templateConfiguration" : {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "createdBy" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "name" : "name",
            "description" : "description",
            "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "secret" : "secret",
            "enabled" : true
          }
        },
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/plugins"
        },
        "properties" : {
          "templateConfiguration" : {
            "$ref" : "#/components/schemas/Plugin_templateConfiguration"
          }
        }
      },
...
      "inline_object_1" : {
        "description" : "Object defining the properties to update and their new values",
        "properties" : {
          "id" : {
            "description" : "The template ID to be installed",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "inline_object_2" : {
        "description" : "Object defining the properties to update and their new values",
        "properties" : {
          "name" : {
            "description" : "The plugin name",
            "type" : "string"
          },
          "description" : {
            "description" : "The plugin description",
            "type" : "string"
          },
          "enabled" : {
            "description" : "The plugin status, enabled or disabled",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "Plugin_templateConfiguration" : {
        "description" : "Main template properties",
        "example" : {
          "createdAt" : "2000-01-23T04:56:07.000+00:00",
          "createdBy" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "name" : "name",
          "description" : "description",
          "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "secret" : "secret",
          "enabled" : true
        },
        "properties" : {
          "name" : {
            "description" : "The plugin name",
            "maxLength" : 256,
            "minLength" : 3,
            "type" : "string"
          },
          "description" : {
            "description" : "The plugin description",
            "maxLength" : 1024,
            "minLength" : 0,
            "type" : "string"
          },
          "createdBy" : {
            "description" : "The ID (Fron Azure AD) of the user that created the plugin",
            "format" : "uuid",
            "type" : "string"
          },
          "createdAt" : {
            "description" : "The date of creation of the plugin",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "description" : "The plugin ID, auto-generated at creation",
            "format" : "uuid",
            "type" : "string"
          },
          "secret" : {
            "description" : "The plugin secret, auto-generated at creation. It may also be regenerated later. Be careful, this property is only available at creation or after an explicit regeneration.\n",
            "format" : "password",
            "type" : "string"
          },
          "enabled" : {
            "default" : true,
            "description" : "The plugin status, enabled or disabled",
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "A valid Microsoft 365 access_token issued by Azure AD",
        "scheme" : "bearer",
        "type" : "http"
      },
      "pluginId" : {
        "description" : "A valid plugin ID",
        "in" : "header",
        "name" : "X-APP-ID",
        "type" : "apiKey"
      },
      "pluginSecret" : {
        "description" : "A valid plugin secret",
        "in" : "header",
        "name" : "X-API-KEY",
        "type" : "apiKey"
      }
    }
  }
}